+++ previous-versions/accounting/src/sdk/models/shared/directcostlineitem.ts
@@ -6,6 +6,7 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { AccountRef } from "./accountref";
 import { ItemRef } from "./itemref";
 import { TaxRateRef } from "./taxrateref";
+import { Tracking } from "./tracking";
 import { TrackingCategoryRef } from "./trackingcategoryref";
 import { Expose, Type } from "class-transformer";
 
@@ -97,7 +98,8 @@ export class DirectCostLineItem extends SpeakeasyBase {
 
     @SpeakeasyMetadata()
     @Expose({ name: "tracking" })
-    tracking?: Record<string, any>;
+    @Type(() => Tracking)
+    tracking?: Tracking;
 
     /**
      * Collection of categories against which this direct cost is tracked.