+++ previous-versions/accounting/src/sdk/models/shared/journalentry.ts
@@ -5,6 +5,7 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { InvoiceTo } from "./invoiceto";
 import { JournalLine } from "./journalline";
+import { JournalRef } from "./journalref";
 import { Metadata } from "./metadata";
 import { SupplementalData } from "./supplementaldata";
 import { Expose, Type } from "class-transformer";
@@ -90,7 +91,8 @@ export class JournalEntry extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "journalRef" })
-    journalRef?: Record<string, any>;
+    @Type(() => JournalRef)
+    journalRef?: JournalRef;
 
     @SpeakeasyMetadata()
     @Expose({ name: "metadata" })