+++ previous-versions/accounting/src/sdk/sdk.ts
@@ -78,9 +78,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.38.0";
-    genVersion = "2.150.0";
-    userAgent = "speakeasy-sdk/typescript 0.38.0 2.150.0 3.0.0 @codat/accounting";
+    sdkVersion = "0.38.1";
+    genVersion = "2.155.1";
+    userAgent = "speakeasy-sdk/typescript 0.38.1 2.155.1 3.0.0 @codat/accounting";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);