+++ previous-versions/accounting/RELEASES.md
@@ -689,3 +689,13 @@ Based on:
 - [typescript v0.38.0] previous-versions/accounting
 ### Releases
 - [NPM v0.38.0] https://www.npmjs.com/package/@codat/accounting/v/0.38.0 - previous-versions/accounting
+
+## 2023-10-16 01:24:39
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.38.1] previous-versions/accounting
+### Releases
+- [NPM v0.38.1] https://www.npmjs.com/package/@codat/accounting/v/0.38.1 - previous-versions/accounting